Bluetooth: ISO: Change ISO seq_num to 16-bit
The sequence number is by the core spec defined as 16-bit.
We had implemented a workaround for the wrapping of the
sequence number, which required the type to be larger
than 16-bit (32-bit).

However, since the definition of the sequence number,
and the use of, is poorly defined by the core spec, we
are reverting this workaround and reducing the sequence
number to 16-bit again. This way it is more in line
with the core spec, as well as more intuitive given the
other uses for the sequence number.

This change moves the responsibility of using the
right value to the upper layers, as the stack can
and will no longer provide any guarantees.

+++ include/zephyr/bluetooth/audio/audio.h
@@ -1935,7 +1935,7 @@ int bt_audio_stream_release(struct bt_audio_stream *stream);
  *  @return Bytes sent in case of success or negative value in case of error.
  */
 int bt_audio_stream_send(struct bt_audio_stream *stream, struct net_buf *buf,
-			 uint32_t seq_num, uint32_t ts);
+			 uint16_t seq_num, uint32_t ts);
 
 /** @brief Parameter struct for the unicast group functions
  *
